remove post analysis error

+++ post_analysis/ICE.sh
@@ -18,6 +18,9 @@ cat $i.output.format.v | awk '{if($5>0){print $2/40000"\t"$4/40000"\t"$5}}' > $i
 contig_name=`head -n 1 $i.output.format.v | cut -f1`
 python ${InfoHiC_lib}/post_analysis/ICE.py $i.output.m
 Rscript ${InfoHiC_lib}/post_analysis/ICE_filter.R $i.output_ice.m $window_size 40000
+if [[ ! -s $i.output_ice.m ]];then
+	exit 1;
+fi
 cat $i.output_ice.m | awk '{if($1<=$2){print "'${contig_name}'\t"($1)*40000"\t'${contig_name}'\t"($2)*40000"\t"$3}}' > $i.output.m.format
 cat $i.output.m.format  |  awk '{if($2>0 && $4>0) print $0}' > $i.output.m.format.v
 Rscript ${InfoHiC_lib}/post_analysis/tad.R $i.output.m.format.v 50
